Re-split ST0 to match structure of other overlays
