+++ src/st/st0/34908.c
@@ -387,126 +387,3 @@ void CheckFieldCollision(u16* hitSensors, s16 sensorCount) {
 INCLUDE_ASM("st/st0/nonmatchings/34908", func_801B5BF0);
 
 #include "../replace_breakable_with_item_drop.h"
-
-void func_801B5EC8(void) {
-    s32 temp_v1;
-    Entity* entity;
-
-    entity = g_CurrentEntity;
-    if (entity->velocityY >= 0) {
-        temp_v1 =
-            entity->ext.generic.unk88.S16.unk0 + entity->ext.generic.unk84.unk;
-        entity->ext.generic.unk84.unk = temp_v1;
-        entity->velocityX = temp_v1;
-        if (temp_v1 == 0x10000 || temp_v1 == -0x10000) {
-            entity->ext.generic.unk88.S16.unk0 =
-                -entity->ext.generic.unk88.S16.unk0;
-        }
-        entity = g_CurrentEntity;
-    }
-
-    if (entity->velocityY < FIX(0.25)) {
-        entity->velocityY += FIX(0.125);
-    }
-}
-
-void func_801B5F4C(u16 arg0) {
-    Collider res;
-
-    if (g_CurrentEntity->velocityX < 0) {
-        g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
-                             g_CurrentEntity->posY.i.hi - 7, &res, 0);
-        if (res.effects & 5) {
-            g_CurrentEntity->velocityY = 0;
-        }
-    }
-
-    g_api.CheckCollision(
-        g_CurrentEntity->posX.i.hi, g_CurrentEntity->posY.i.hi + 7, &res, 0);
-
-    if (arg0) {
-        if (!(res.effects & 5)) {
-            MoveEntity();
-            FallEntity();
-            return;
-        }
-
-        g_CurrentEntity->velocityX = 0;
-        g_CurrentEntity->velocityY = 0;
-
-        if (res.effects & 4) {
-            g_CurrentEntity->posY.val += FIX(0.125);
-            return;
-        }
-
-        g_CurrentEntity->posY.i.hi += res.unk18;
-        return;
-    }
-
-    if (!(res.effects & 5)) {
-        MoveEntity();
-        func_801B5EC8();
-    }
-}
-
-// Can't be deduped with the other instances due to the extra memcpy
-void CollectHeart(u16 index) {
-    s8 hearts[10];
-
-    __builtin_memcpy(&hearts[0], &c_HeartPrizes[0], 2);
-    __builtin_memcpy(&hearts[8], &c_HeartPrizes[4], 2);
-    g_api.PlaySfx(SFX_HEART_PICKUP);
-    g_Status.hearts = hearts[index] + g_Status.hearts;
-    if (g_Status.heartsMax < g_Status.hearts) {
-        g_Status.hearts = g_Status.heartsMax;
-    }
-    DestroyEntity(g_CurrentEntity);
-}
-
-#include "../collect_gold.h"
-
-void func_801B61D4(void) {}
-
-void UnusedDestroyCurrentEntity(void) { DestroyEntity(g_CurrentEntity); }
-
-void CollectSubweapon(u16 subWeaponIdx) {
-    Entity* player = &PLAYER;
-    u16 subWeapon;
-
-    g_api.PlaySfx(SFX_ITEM_PICKUP);
-    subWeapon = g_Status.subWeapon;
-    g_Status.subWeapon = D_80181CAC[subWeaponIdx];
-
-    if (subWeapon == g_Status.subWeapon) {
-        subWeapon = 1;
-        g_CurrentEntity->unk6D[0] = 0x10;
-    } else {
-        subWeapon = D_80181CDC[subWeapon];
-        g_CurrentEntity->unk6D[0] = 0x60;
-    }
-
-    if (subWeapon != 0) {
-        g_CurrentEntity->params = subWeapon;
-        g_CurrentEntity->posY.i.hi = player->posY.i.hi + 12;
-        g_CurrentEntity->step = 7;
-        g_CurrentEntity->step_s = 0;
-        g_CurrentEntity->velocityY = FIX(-2.5);
-        g_CurrentEntity->animCurFrame = 0;
-        g_CurrentEntity->ext.generic.unk88.S16.unk2 = 5;
-        if (player->facingLeft != 1) {
-            g_CurrentEntity->velocityX = FIX(-2);
-            return;
-        }
-        g_CurrentEntity->velocityX = FIX(2);
-        return;
-    }
-    DestroyEntity(g_CurrentEntity);
-}
-
-void DestroyCurrentEntity(void) { DestroyEntity(g_CurrentEntity); }
-
-Entity* func_801B633C(void) {
-    g_CurrentEntity->step = 3;
-    g_CurrentEntity->params = 4;
-    return g_CurrentEntity;
-}
